Add files via upload
This work aimed to develop an affordable trans-radial prosthetic arm. 8-channel, 1 KHz sEMG data from 35 subjects using the MYO Thalmic bracelet was utilised to demonstrate the proof of concept. The signals were denoised using a 50 - 450 Hz bandpass filter. Subsequently, the signal was normalized and transformed using PCA. Two classifiers, SVM and RFC, were employed for the analysis, with SVM exhibiting the best performance. It achieved a mean classification accuracy of 88.87% for six gestures. This accuracy can be further improved by implementing suitable post-processing techniques such as Bayesian fusion or joint-window decision-making. While the analysis was conducted in Python, the availability of C++ alternatives for most libraries allows for a relatively straightforward porting of this solution.